Fix: 캐릭터의 보스 공격 구현 수
-보스 체력을 플레이어의 콤보 공격 세기에 따라서 깎기
-보스 공격 성공시 캐릭터 해킹 포인트 +1 오르기
-boss_hp가 0이 되고 hacking_point를 전부 쌓으면 임무완료(MissionClear) 씬으로 넘어가기

+++ EPOH/Assets/Kyoungmin/Scripts/BossManager.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class BossManager : MonoBehaviour
 {
-    public int boss_hp = 1000;
-    public int player_hp = 200;
-    public int hacking_point = 0;
+    public float boss_hp = 1000f;
+    public float player_hp = 200f;
+    public float hacking_point = 0f;
 }
